+++ wp-content/themes/nolan-young-demo-theme-003/page-templates/page-template-ppc-lp-2026.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Template Name: PPC Landing Page 2026
+ *
+ * @package NolanYoungDemoTheme003
+ */
+
+get_header();
+?>
+<main id="content">
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-hero' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-sect01' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-sect02' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-sect03' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-sect04' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-sect05' ); ?>
+	<?php get_template_part( 'template-parts/content', 'ppc-lp-2026-cta' ); ?>
+</main>
+<?php get_footer();